+++ glide/estimators/ptd.py
@@ -1,16 +1,17 @@
+from math import floor
 from typing import Optional, Tuple
 
 import numpy as np
 from numpy.typing import NDArray
 
 from glide.confidence_intervals import BootstrapConfidenceInterval
+from glide.estimators.classical import ClassicalMeanEstimator
 from glide.estimators.ptd_core import (
     _compute_bootstrap_labeled_means,
     _compute_bootstrap_mean_estimates,
     _compute_tuning_parameter,
 )
 from glide.mean_inference_results import PredictionPoweredMeanInferenceResult
-from glide.utils import compute_effective_sample_size
 
 
 class PTDMeanEstimator:
@@ -155,7 +156,8 @@ class PTDMeanEstimator:
             bootstrap_estimates=bootstrap_mean_estimates,
             confidence_level=confidence_level,
         )
-        effective_sample_size = compute_effective_sample_size(y_true_filtered, confidence_interval.var)
+        classical_confidence_interval = ClassicalMeanEstimator().estimate(y_true_filtered).confidence_interval
+        effective_sample_size = floor(n_labeled * classical_confidence_interval.std**2 / confidence_interval.var)
         result = PredictionPoweredMeanInferenceResult(
             confidence_interval=confidence_interval,
             metric_name=metric_name,